requirements.txt
flask>=3.0
torch>=2.0
torchvision>=0.15
Pillow>=10.0
scikit-learn>=1.3
